backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Educational video generator backend with Peter Griffin voice"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.1",
    "google-generativeai>=0.8.5",
    "elevenlabs>=1.0.0",
    "assemblyai>=0.15.0",
    "aiofiles>=23.2.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.metadata]
allow-direct-references = true
